requirements.txt
streamlit>=1.34.0
requests
opencage
pandas
numpy
pydeck
altair
pillow
openai
regex
keras
tensorflow
scikit-learn